ADD: Money object
The plan is to add a money object into a bank object, thereby encapsulating it from the bank,
my thought is that the money may in future have a seperate owner and need to be different from the
bank. So far i have applied simplecov to ensure test coverage, and created and passed the first test
which creates and checks a money object.