ADD create transaction method on the account object.
Adding test and code for the create transaction method on the account object. Currently takes
three arguments, for amount, deposit and a default time object i.e. the time the transaction
is created which can be formatted later. Think i may refactor the code to encapsulate the
transaction away from the account now.