REFACTOR: remove most of the arguments in creating a transaction.
We dont need and probably dont want to allow a user to specify whether an amount is a
contribution or withdrawal as its kind of confusing, so am removing the option. Adding a
simple if loop to define deposit and withdrawal else both are false, also setting the date/time
at the moment of the transaction, as it seems like a better idea then allowing the user the
opiton of setting it.