ADD: adding and withdrawing from the account balance.
Ability to add and deduct from the account balance, and a starting balance of zero. Not sure
if the code is especially tidy, but function is working. Experimented with accessor methods,
but i dont think that is right in this instance as may need the methods for use in printing
statement. Still got some work to do.