ADD: date method and formatting of that with a mock of the date on the test.
quite a nice test driven section, at initialization of the object the object is
given a date, that date is set as a default so always created at inception and
have applied a method on the class via the test to call that date. Clearly a
date is always going to change, so have had to used a gem called timecop to make it
work... which mocks the date so test will pass. Have also formatted the date in
the return from date method.