REFACTOR mocking date, credit/debit ability and more
playing with the code and user stories led me to create withdrawal
and deposit methods for adding to the account, and withdrawaling,
but that was then showing as positive amounts on the statement and
transactons plus the acceptance criteria requires the values to be
on seperate columns so have applied a credit / debit attribute so
the values can be dealt with seperatly.

Also formatted the statement body, which now shows as floats, so am
close to hitting acceptance criteria. Test coverage is high, but am
not obsessed by getting 100% coverage, may review again shortly for
more tests.