TESTING: pushing to 100% coverage.
Have completed i decided to get the test coverage up, so covered
missing lines and uncovered a bug in the statement where withdrawals
were not being deducted from the statement, exactly why tests are
so great! Have fixed that and now have 100% coverage.

Given the statement appears to work and i have decent test coverage
and i think that i will stop with this exercise for now.